dc: Remap DC types <--> Java primitives
* Signed ints are now their Java equivalent
* uint64 is now stored in a long

+++ src/main/java/astron/dc/DCTokens.java
@@ -33,14 +33,14 @@ public class DCTokens {
 
         DATA_CLASSES.put(DDataTypes.STRING, String.class);
         DATA_CLASSES.put(DDataTypes.BOOL, Boolean.class);
-        DATA_CLASSES.put(DDataTypes.INT8, Integer.class);
-        DATA_CLASSES.put(DDataTypes.INT16, Integer.class);
+        DATA_CLASSES.put(DDataTypes.INT8, Byte.class);
+        DATA_CLASSES.put(DDataTypes.INT16, Short.class);
         DATA_CLASSES.put(DDataTypes.INT32, Integer.class);
-        DATA_CLASSES.put(DDataTypes.INT64, Integer.class);
+        DATA_CLASSES.put(DDataTypes.INT64, Long.class);
         DATA_CLASSES.put(DDataTypes.UINT8, Integer.class);
         DATA_CLASSES.put(DDataTypes.UINT16, Integer.class);
         DATA_CLASSES.put(DDataTypes.UINT32, Integer.class);
-        DATA_CLASSES.put(DDataTypes.UINT64, Integer.class);
+        DATA_CLASSES.put(DDataTypes.UINT64, Long.class);
 
     }
